backend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
ecdsa==0.17.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
jose==1.0.0
mysql-connector-python==8.0.29
passlib==1.7.4
protobuf==4.21.2
pyasn1==0.4.8
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
python-jose==3.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing==3.7.4.3
typing_extensions==4.2.0
uvicorn==0.17.6
